Create V1
This model is purely based on Hill functions to model the dose response behaviour of inducible promoters in a metabolic networks and gene expression repression. This script does not execute dynamical simulations, rather seeks to identify the steady state answer for a range of inducer concentrations - of which an intermediate conc combination of inducers is strongest.